+++ services/cdn/src/stackit/cdn/models/http_backend.py
@@ -30,11 +30,11 @@ class HttpBackend(BaseModel):
         description="An object mapping multiple alternative origins to country codes.  Any request from one of those country codes will route to the alternative origin. Do note that country codes may only be used once. You cannot have a country be assigned to multiple alternative origins. "
     )
     origin_request_headers: Dict[str, StrictStr] = Field(
-        description="Headers that will be sent with every request to the configured origin.  **WARNING**: Do not store sensitive values in the headers.  The configuration is stored as plain text. ",
+        description="Headers that will be sent with every request to the configured origin.  **WARNING**: Do not store sensitive values in the headers. The configuration is stored as plain text. ",
         alias="originRequestHeaders",
     )
     origin_url: StrictStr = Field(
-        description="The origin of the content that should be made available through the CDN.   Note that the path and query parameters are ignored. Ports are allowed. If no protocol is provided, `https` is assumed.   So `www.example.com:1234/somePath?q=123` is normalized to `https://www.example.com:1234` ",
+        description="The origin of the content that should be made available through the CDN. Note that the path and query parameters are ignored. Ports are allowed. If no protocol is provided, `https` is assumed. So `www.example.com:1234/somePath?q=123` is normalized to `https://www.example.com:1234` ",
         alias="originUrl",
     )
     type: StrictStr