+++ services/cdn/src/stackit/cdn/models/waf_paranoia_level.py
@@ -21,7 +21,7 @@ from typing_extensions import Self
 
 class WafParanoiaLevel(str, Enum):
     """
-    The paranoia level defines how aggressively the WAF should action on requests.   It ranges from `L1` (least strict, lowest chance of false positives) to `L4` (most strict, highest chance of false positives). A higher paranoia level is more effective at catching attacks but can also block legitimate traffic.
+    The paranoia level defines how aggressively the WAF should action on requests. It ranges from `L1` (least strict, lowest chance of false positives) to `L4` (most strict, highest chance of false positives). A higher paranoia level is more effective at catching attacks but can also block legitimate traffic.
     """
 
     """