+++ services/cdn/CHANGELOG.md
@@ -1,3 +1,10 @@
+## v2.5.0
+- **Feature:** New model classes `TlsConfigPatch` and `TlsConfig`
+- **Feature:** add attributes `forwardHostHeader`, `stripResponseCookies` and `tls` to the model classes:
+  - `Config`
+  - `ConfigPatch`
+  - `CreateDistributionPayload`
+
 ## v2.4.0
 - **Feature:** client now supports UUID and decimal types
 - **Bugfix:** timeouts now passed to requests library